requirements.txt
flask
flask-restful
mysql-connector-python
Flask-SQLAlchemy
flask_httpauth
python-dotenv
pytest